Moving Typescript definitions around a bit.
